Refactor Application edit
