+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store-new/source/src/app/components/Applications/Edit/EditApp.jsx
@@ -27,11 +27,12 @@ import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
 import CloseIcon from '@material-ui/icons/Close';
 import Slide from '@material-ui/core/Slide';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, injectIntl } from 'react-intl';
+import { Link } from 'react-router-dom';
 import API from 'AppData/api';
 import ApplicationCreateForm from 'AppComponents/Shared/AppsAndKeys/ApplicationCreateForm';
 import Alert from 'AppComponents/Shared/Alert';
-import { ScopeValidation, resourceMethods, resourcePaths } from 'AppComponents/Shared/ScopeValidation';
+import Application from '../../../data/Application';
 
 /**
  *
@@ -44,11 +45,18 @@ const styles = theme => ({
         backgroundColor: theme.palette.background.appBar,
         color: theme.palette.getContrastText(theme.palette.background.appBar),
     },
+    buttonRight: {
+        textDecoration: 'none',
+        color: 'white',
+    },
     flex: {
         flex: 1,
     },
     button: {
         marginRight: theme.spacing.unit * 2,
+    },
+    link: {
+        textDecoration: 'none',
     },
     buttonWrapper: {
         paddingLeft: theme.spacing.unit * 7,
@@ -65,20 +73,21 @@ function Transition(props) {
     return <Slide direction='up' {...props} />;
 }
 /**
- * Component used to handle application creation
- * @class NewApp
+ * Component used to handle application editing
+ * @class EditApp
  * @extends {React.Component}
  * @param {any} value @inheritDoc
  */
-class NewApp extends React.Component {
+class EditApp extends React.Component {
     /**
      * @param {*} props properties
      */
     constructor(props) {
         super(props);
         this.state = {
-            open: false,
+            open: true,
             applicationRequest: {
+                applicationId: null,
                 name: '',
                 throttlingPolicy: '',
                 description: '',
@@ -94,24 +103,30 @@ class NewApp extends React.Component {
     /**
      * Get all the throttling Policies from backend and
      * update the state
-     * @memberof NewApp
+     * @memberof EditApp
      */
     componentDidMount() {
+        const { match } = this.props;
+        const { applicationRequest } = this.state;
+        const promisedApplication = Application.get(match.params.application_id);
         // Get all the tires to populate the drop down.
         const api = new API();
         const promiseTiers = api.getAllTiers('application');
         const promisedAttributes = api.getAllApplicationAttributes();
-        Promise.all([promiseTiers, promisedAttributes])
+        Promise.all([promisedApplication, promiseTiers, promisedAttributes])
             .then((response) => {
-                const [tierResponse, allAttributes] = response;
-                const { applicationRequest } = this.state;
-                const throttlingPolicyList = tierResponse.body.list.map(item => item.name);
-                const newRequest = { ...applicationRequest };
-                if (throttlingPolicyList.length > 0) {
-                    [newRequest.throttlingPolicy] = throttlingPolicyList;
-                }
+                const [application, tierResponse, allAttributes] = response;
+                const throttlingPolicyList = [];
+                tierResponse.body.list.map(item => throttlingPolicyList.push(item.name));
                 const allAppAttributes = [];
                 allAttributes.body.list.map(item => allAppAttributes.push(item));
+                const newRequest = { ...applicationRequest };
+                newRequest.applicationId = application.applicationId;
+                newRequest.name = application.name;
+                newRequest.throttlingPolicy = application.throttlingPolicy;
+                newRequest.description = application.description;
+                newRequest.tokenType = application.tokenType;
+                newRequest.attributes = application.attributes;
                 this.setState({ applicationRequest: newRequest, throttlingPolicyList, allAppAttributes });
             })
             .catch((error) => {
@@ -138,7 +153,7 @@ class NewApp extends React.Component {
     /**
      * @param {object} name application attribute name
      * @returns {void}
-     * @memberof NewApp
+     * @memberof EditApp
      */
     handleAttributesChange = name => (event) => {
         const { applicationRequest } = this.state;
@@ -149,7 +164,7 @@ class NewApp extends React.Component {
     /**
      * @param {object} name application attribute name
      * @returns {Object} attribute value
-     * @memberof NewApp
+     * @memberof EditApp
      */
     getAttributeValue = (name) => {
         const { applicationRequest } = this.state;
@@ -159,7 +174,7 @@ class NewApp extends React.Component {
     /**
      * @param {object} name application attribute name
      * @returns {void}
-     * @memberof NewApp
+     * @memberof EditApp
      */
     isRequiredAttribute = (name) => {
         const { allAppAttributes } = this.state;
@@ -174,42 +189,46 @@ class NewApp extends React.Component {
     };
 
     /**
-     * Validate and send the application create
-     * request to the backend
-     * @memberof NewApp
+     * @param {Object} event the event object
+     * @memberof EditApp
      */
-    saveApplication = () => {
+    handleSubmit = () => {
         const { applicationRequest } = this.state;
-        const { updateApps } = this.props;
+        const {
+            history, intl,
+        } = this.props;
+        this.validateName(applicationRequest.name);
         const api = new API();
-        this.validateName(applicationRequest.name)
-            .then(() => api.createApplication(applicationRequest))
-            .then(() => {
-                console.log('Application created successfully.');
-                this.setState({ open: false });
-                updateApps();
+        const promisedUpdate = api.updateApplication(applicationRequest, null);
+        promisedUpdate
+            .then((response) => {
+                const appId = response.body.applicationId;
+                const redirectUrl = '/applications/' + appId;
+                history.push(redirectUrl);
+                Alert.info(intl.formatMessage({
+                    id: 'Applications.Edit.app.updated.success',
+                    defaultMessage: 'Application updated successfully',
+                }));
+                console.log('Application updated successfully.');
             })
             .catch((error) => {
-                const { response } = error;
-                if (response && response.body) {
-                    const message = response.body.description || 'Error while creating the application';
-                    Alert.error(message);
-                } else {
-                    Alert.error(error.message);
-                }
-                console.error('Error while creating the application');
+                Alert.error(intl.formatMessage({
+                    id: 'Applications.Edit.error.update.app',
+                    defaultMessage: 'Error while updating application',
+                }));
+                console.log('Error while updating application ' + error);
             });
     };
 
     /**
-     * @memberof NewApp
+     * @memberof EditApp
      */
     handleClose = () => {
         this.setState({ open: false });
     };
 
     /**
-     * @memberof NewApp
+     * @memberof EditApp
      */
     handleClickOpen = () => {
         this.setState({ open: true });
@@ -226,7 +245,7 @@ class NewApp extends React.Component {
 
     /**
      * @inheritdoc
-     * @memberof NewApp
+     * @memberof EditApp
      */
     render() {
         const {
@@ -235,37 +254,17 @@ class NewApp extends React.Component {
         const { classes } = this.props;
         return (
             <React.Fragment>
-                <ScopeValidation resourcePath={resourcePaths.APPLICATIONS} resourceMethod={resourceMethods.POST}>
-                    <Button
-                        variant='contained'
-                        color='primary'
-                        className={classes.button}
-                        onClick={this.handleClickOpen}
-                    >
-                        <FormattedMessage
-                            defaultMessage='ADD NEW APPLICATION'
-                            id='Applications.Create.NewApp.add.new.application.button'
-                        />
-                    </Button>
-                </ScopeValidation>
                 <Dialog fullScreen open={open} onClose={this.handleClose} TransitionComponent={Transition}>
                     <AppBar className={classes.appBar}>
                         <Toolbar>
-                            <IconButton color='inherit' onClick={this.handleClose} aria-label='Close'>
-                                <CloseIcon />
-                            </IconButton>
+                            <Link to='/applications' className={classes.buttonRight}>
+                                <IconButton color='inherit' onClick={this.handleClose} aria-label='Close'>
+                                    <CloseIcon />
+                                </IconButton>
+                            </Link>
                             <Typography variant='title' color='inherit' className={classes.flex}>
-                                <FormattedMessage
-                                    defaultMessage='Create New Application'
-                                    id='Applications.Create.NewApp.create.new.application.title'
-                                />
+                                <FormattedMessage id='Applications.Edit.edit.app' defaultMessage='Edit Application' />
                             </Typography>
-                            <Button color='inherit' onClick={this.handleClose}>
-                                <FormattedMessage
-                                    defaultMessage='Save'
-                                    id='Applications.Create.NewApp.save.application'
-                                />
-                            </Button>
                         </Toolbar>
                     </AppBar>
                     <div className={classes.createFormWrapper}>
@@ -282,21 +281,26 @@ class NewApp extends React.Component {
                         />
                     </div>
                     <div className={classes.buttonWrapper}>
-                        <Button variant='outlined' className={classes.button} onClick={this.handleClose}>
-                            <FormattedMessage
-                                defaultMessage='Cancel'
-                                id='Applications.Create.NewApp.cancel'
-                            />
-                        </Button>
+                        <Link to='/applications' className={classes.link}>
+                            <Button
+                                variant='outlined'
+                                className={classes.button}
+                            >
+                                <FormattedMessage
+                                    id='Applications.Edit.cancel'
+                                    defaultMessage='CANCEL'
+                                />
+                            </Button>
+                        </Link>
                         <Button
                             variant='contained'
-                            color='primary'
                             className={classes.button}
-                            onClick={this.saveApplication}
+                            color='primary'
+                            onClick={this.handleSubmit}
                         >
                             <FormattedMessage
-                                defaultMessage='ADD NEW APPLICATION'
-                                id='Applications.Create.NewApp.add.new.application.create'
+                                id='Applications.Edit.update.app'
+                                defaultMessage='UPDATE APPLICATION'
                             />
                         </Button>
                     </div>
@@ -306,8 +310,9 @@ class NewApp extends React.Component {
     }
 }
 
-NewApp.propTypes = {
+EditApp.propTypes = {
     classes: PropTypes.shape({}).isRequired,
+    intl: PropTypes.shape({}).isRequired,
 };
 
-export default withStyles(styles)(NewApp);
+export default injectIntl(withStyles(styles)(EditApp));
