+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store-new/source/src/app/components/Shared/Alert.jsx
@@ -53,7 +53,6 @@ class Alert {
     show() {
         const promisedInstance = this._getMessageInstance();
         const { onClose } = this;
-        const { intl } = this.prop;
         promisedInstance
             .then((instance) => {
                 instance.notice({
@@ -68,8 +67,7 @@ class Alert {
                     ),
                 });
             })
-            .catch(error => console.error(intl.formatMessage({
-                defaultMessage: 'Error while showing alert'}) + error));
+            .catch(error => console.error('Error while showing alert' + error));
         /* TODO: Remove above console error with logging library error method */
     }
